Move auth helpers to a more central location
